+++ backend/src/main/java/groom/backend/domain/opendata/listener/ChargerWarmupListener.java
@@ -1,6 +1,7 @@
 package groom.backend.domain.opendata.listener;
 
 import groom.backend.domain.opendata.dto.response.ChargerLocationResponse;
+import groom.backend.domain.opendata.service.spec.ChargerLocationFindService;
 import groom.backend.domain.opendata.service.spec.ChargerLocationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,6 +16,7 @@ import java.util.List;
 @Slf4j
 public class ChargerWarmupListener {
   private final ChargerLocationService chargerLocationService;
+  private final ChargerLocationFindService chargerLocationFindService;
 
   @EventListener
   public void listen(ApplicationReadyEvent event) {
@@ -24,7 +26,7 @@ public class ChargerWarmupListener {
             .toList();
     log.info("총 충전소 개수: {}", ids.size());
     ids.stream()
-            .map(chargerLocationService::getChargerLocationById) // 여기서 캐시 활용
+            .map(chargerLocationFindService::getChargerLocationById) // 여기서 캐시 활용
             .toList();
   }
 }
